pyproject.toml
[project]
name = "voice-assistant-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.0",
    "ollama>=0.6.1",
    "onnxruntime>=1.23.2",
    "piper-tts>=1.3.0",
    "pvporcupine>=4.0.1",
    "sounddevice>=0.5.3",
    "tokenizers>=0.22.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
